Added major change
Added changes to reset box comm scheduler.
The reset box sends command to the DAC units and fetches the information.
If multiple reset boxes are connected to multiple DACs, a single reset box will be master(meaning, sending the commands) and the other reset box(s) will be listeners.
It is tested with 2 reset boxes and any moment of turning on, the reset units syncd up. One minor issue is with the listener(sometimes, the 80 bytes reception goes wrong but gets corrected in the next cycle)